+++ application/package.json
@@ -2,7 +2,7 @@
   "name": "frontend",
   "version": "0.1.0",
   "private": true,
-  "proxy": "http://localhost:3001/",
+  "proxy": "http://localhost:3001/api",
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
