+++ application/server/routes/pages/login.js
@@ -1,9 +1,9 @@
 const express = require("express");
 const router = express.Router();
-const { user } = require("../controller");
+const { user } = require("../../controller");
 
 router.post("/", (req, res) => {
   user.login(req, res);
-})
+});
 
 module.exports = router;
