+++ application/server/routes/pages/register.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const { user } = require("../controller");
+const { user } = require("../../controller");
 
 router.post("/", (req, res) => {
   user.newUser(req, res);
