+++ application/server/routes/pages/root.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const status = require("../status");
+const status = require("../../status");
 const router = express.Router();
 
 router.get("/", (req, res) => {
