--- application/server/routes/upload.js
@@ -1,10 +1,0 @@
-const express = require('express');
-const isAuth = require("../middleware/is-auth");
-const status = require('../status');
-const router = express.Router();
-
-router.get('/', isAuth, (req, res) => {
-  status.Ok(req, res);
-})
-
-module.exports = router;
