+++ application/server/routes/utilities/session-check.js
@@ -1,6 +1,6 @@
 const express = require("express");
-const isAuth = require("../middleware/is-auth");
-const status = require("../status");
+const isAuth = require("../../middleware/is-auth");
+const status = require("../../status");
 const router = express.Router();
 
 router.get("/", isAuth, (req, res) => {
