+++ application/src/components/utilities/character-count-display.js
