+++ application/src/components/utilities/format-number.js
